backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-filter==24.2
djangorestframework-simplejwt==5.3.1
psycopg[binary,pool]==3.2.4
dj-database-url==2.2.0
drf-spectacular==0.27.2
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.0
python-dotenv==1.0.1
gunicorn==22.0.0
whitenoise==6.7.0
django-cors-headers==4.4.0

